refactor: extract submodules to satisfy SPEC_0021 file-size limit
- emit.rs: extract interpreter fallback to emit/interpreter.rs (2134→1676 lines)
- render_solve.rs: extract per-dialect op renderers to render_solve_ops.rs (2115→1886 lines)
- lower.rs: extract builtin dispatch to lower/builtin_dispatch.rs (2148→1922 lines)
