+++ crates/rumoca-phase-solve/src/lower/builtin_dispatch.rs
@@ -1,0 +1,230 @@
+use super::*;
+
+impl<'a> LowerBuilder<'a> {
+    fn lower_simple_builtin(
+        &mut self,
+        function: rumoca_core::BuiltinFunction,
+        args: &[rumoca_core::Expression],
+        scope: &Scope,
+        call_depth: usize,
+        span: rumoca_core::Span,
+    ) -> Option<Result<Reg, LowerError>> {
+        let arg = |builder: &mut Self, idx: usize| -> Result<Reg, LowerError> {
+            let expr = args.get(idx).ok_or_else(|| LowerError::ContractViolation {
+                reason: format!(
+                    "builtin {:?} requires argument at index {}, but only {} were provided",
+                    function,
+                    idx,
+                    args.len()
+                ),
+                span: rumoca_core::Span::DUMMY,
+            })?;
+            builder.lower_expr(expr, scope, call_depth)
+        };
+        let unary = |builder: &mut Self, op: UnaryOp| -> Result<Reg, LowerError> {
+            let x = arg(builder, 0)?;
+            Ok(builder.emit_unary(op, x))
+        };
+        let binary = |builder: &mut Self, op: BinaryOp| -> Result<Reg, LowerError> {
+            let x = arg(builder, 0)?;
+            let y = arg(builder, 1)?;
+            Ok(builder.emit_binary(op, x, y))
+        };
+
+        let result = match function {
+            rumoca_core::BuiltinFunction::Abs => unary(self, UnaryOp::Abs),
+            rumoca_core::BuiltinFunction::Sign => unary(self, UnaryOp::Sign),
+            rumoca_core::BuiltinFunction::Sqrt => unary(self, UnaryOp::Sqrt),
+            rumoca_core::BuiltinFunction::Floor | rumoca_core::BuiltinFunction::Integer => {
+                unary(self, UnaryOp::Floor)
+            }
+            rumoca_core::BuiltinFunction::Ceil => unary(self, UnaryOp::Ceil),
+            rumoca_core::BuiltinFunction::Sin => unary(self, UnaryOp::Sin),
+            rumoca_core::BuiltinFunction::Cos => unary(self, UnaryOp::Cos),
+            rumoca_core::BuiltinFunction::Tan => unary(self, UnaryOp::Tan),
+            rumoca_core::BuiltinFunction::Asin => unary(self, UnaryOp::Asin),
+            rumoca_core::BuiltinFunction::Acos => unary(self, UnaryOp::Acos),
+            rumoca_core::BuiltinFunction::Atan => unary(self, UnaryOp::Atan),
+            rumoca_core::BuiltinFunction::Sinh => unary(self, UnaryOp::Sinh),
+            rumoca_core::BuiltinFunction::Cosh => unary(self, UnaryOp::Cosh),
+            rumoca_core::BuiltinFunction::Tanh => unary(self, UnaryOp::Tanh),
+            rumoca_core::BuiltinFunction::Exp => unary(self, UnaryOp::Exp),
+            rumoca_core::BuiltinFunction::Log => unary(self, UnaryOp::Log),
+            rumoca_core::BuiltinFunction::Log10 => unary(self, UnaryOp::Log10),
+            rumoca_core::BuiltinFunction::Atan2 => binary(self, BinaryOp::Atan2),
+            rumoca_core::BuiltinFunction::Div => self.lower_div_builtin(args, scope, call_depth),
+            rumoca_core::BuiltinFunction::Mod => {
+                self.lower_remainder_builtin(args, scope, call_depth, UnaryOp::Floor, "mod")
+            }
+            rumoca_core::BuiltinFunction::Rem => {
+                self.lower_remainder_builtin(args, scope, call_depth, UnaryOp::Trunc, "rem")
+            }
+            rumoca_core::BuiltinFunction::NoEvent => arg(self, 0),
+            rumoca_core::BuiltinFunction::Homotopy => {
+                arg(self, usize::from(self.is_initial_mode && args.len() > 1))
+            }
+            rumoca_core::BuiltinFunction::Smooth => arg(self, 1),
+            rumoca_core::BuiltinFunction::Zeros => Ok(self.emit_const(0.0)),
+            rumoca_core::BuiltinFunction::Ones => Ok(self.emit_const(1.0)),
+            rumoca_core::BuiltinFunction::Fill
+            | rumoca_core::BuiltinFunction::Scalar
+            | rumoca_core::BuiltinFunction::Vector
+            | rumoca_core::BuiltinFunction::Matrix
+            | rumoca_core::BuiltinFunction::Diagonal
+            | rumoca_core::BuiltinFunction::Transpose
+            | rumoca_core::BuiltinFunction::Linspace
+            | rumoca_core::BuiltinFunction::Cat
+            | rumoca_core::BuiltinFunction::Cross
+            | rumoca_core::BuiltinFunction::Skew
+            | rumoca_core::BuiltinFunction::OuterProduct
+            | rumoca_core::BuiltinFunction::Symmetric => {
+                self.lower_builtin_first_array_like_value(function, args, scope, call_depth, span)
+            }
+            rumoca_core::BuiltinFunction::Ndims => {
+                let dims = match self.infer_expr_dims(
+                    args.first()
+                        .expect("ndims() requires exactly 1 argument — malformed Solve-IR"),
+                    scope,
+                ) {
+                    Ok(dims) => dims,
+                    Err(err) => return Some(Err(err)),
+                };
+                Ok(self.emit_const(dims.len() as f64))
+            }
+            rumoca_core::BuiltinFunction::Identity => Ok(self.emit_const(1.0)),
+            _ => return None,
+        };
+        Some(result)
+    }
+
+    fn lower_div_builtin(
+        &mut self,
+        args: &[rumoca_core::Expression],
+        scope: &Scope,
+        call_depth: usize,
+    ) -> Result<Reg, LowerError> {
+        if args.len() != 2 {
+            return Err(LowerError::ContractViolation {
+                reason: format!("div() requires exactly 2 arguments, got {}", args.len()),
+                span: rumoca_core::Span::DUMMY,
+            });
+        }
+        let x = self.lower_expr(&args[0], scope, call_depth)?;
+        let y = self.lower_expr(&args[1], scope, call_depth)?;
+        let q = self.emit_binary(BinaryOp::Div, x, y);
+        Ok(self.emit_unary(UnaryOp::Trunc, q))
+    }
+
+    fn lower_remainder_builtin(
+        &mut self,
+        args: &[rumoca_core::Expression],
+        scope: &Scope,
+        call_depth: usize,
+        quotient_rounding: UnaryOp,
+        function_name: &str,
+    ) -> Result<Reg, LowerError> {
+        if args.len() != 2 {
+            return Err(LowerError::ContractViolation {
+                reason: format!(
+                    "{function_name}() requires exactly 2 arguments, got {}",
+                    args.len()
+                ),
+                span: rumoca_core::Span::DUMMY,
+            });
+        }
+        let x = self.lower_expr(&args[0], scope, call_depth)?;
+        let y = self.lower_expr(&args[1], scope, call_depth)?;
+        let q = self.emit_binary(BinaryOp::Div, x, y);
+        let q_rounded = self.emit_unary(quotient_rounding, q);
+        let product = self.emit_binary(BinaryOp::Mul, q_rounded, y);
+        Ok(self.emit_binary(BinaryOp::Sub, x, product))
+    }
+
+    pub(super) fn lower_builtin(
+        &mut self,
+        function: rumoca_core::BuiltinFunction,
+        args: &[rumoca_core::Expression],
+        span: rumoca_core::Span,
+        scope: &Scope,
+        call_depth: usize,
+    ) -> Result<Reg, LowerError> {
+        if let Some(result) = self.lower_simple_builtin(function, args, scope, call_depth, span) {
+            return result;
+        }
+
+        let arg = |builder: &mut Self, idx: usize| -> Result<Reg, LowerError> {
+            let expr = args.get(idx).ok_or_else(|| LowerError::ContractViolation {
+                reason: format!(
+                    "builtin {:?} requires argument at index {}, but only {} were provided",
+                    function,
+                    idx,
+                    args.len()
+                ),
+                span,
+            })?;
+            builder.lower_expr(expr, scope, call_depth)
+        };
+
+        match function {
+            rumoca_core::BuiltinFunction::Min => {
+                self.lower_min_max_builtin(args, scope, call_depth, BinaryOp::Min, f64::INFINITY)
+            }
+            rumoca_core::BuiltinFunction::Max => self.lower_min_max_builtin(
+                args,
+                scope,
+                call_depth,
+                BinaryOp::Max,
+                f64::NEG_INFINITY,
+            ),
+            // MLS §3.7.2: delay(expr, delayTime) reads expr at a previous
+            // time instant. During event iteration it must not feed the
+            // current unknown value back into the same discrete solve.
+            rumoca_core::BuiltinFunction::Delay => {
+                if args.is_empty() {
+                    return Err(LowerError::ContractViolation {
+                        reason: "delay() requires at least 1 argument".to_string(),
+                        span,
+                    });
+                }
+                self.lower_expr_in_mode(&args[0], scope, call_depth, ValueMode::Pre)
+            }
+            rumoca_core::BuiltinFunction::SemiLinear => {
+                let x = arg(self, 0)?;
+                let k1 = arg(self, 1)?;
+                let k2 = arg(self, 2)?;
+                let zero = self.emit_const(0.0);
+                let cond = self.emit_compare(CompareOp::Ge, x, zero);
+                let pos = self.emit_binary(BinaryOp::Mul, k1, x);
+                let neg = self.emit_binary(BinaryOp::Mul, k2, x);
+                Ok(self.emit_select(cond, pos, neg))
+            }
+            rumoca_core::BuiltinFunction::Der => Ok(self.emit_const(0.0)),
+            rumoca_core::BuiltinFunction::Pre => Err(unsupported_at(
+                "pre() must be lowered to __pre__ parameters before Solve-IR lowering",
+                span,
+            )),
+            rumoca_core::BuiltinFunction::Edge => self.lower_edge_builtin(args, scope, call_depth),
+            rumoca_core::BuiltinFunction::Change => {
+                self.lower_change_builtin(args, scope, call_depth)
+            }
+            rumoca_core::BuiltinFunction::Initial => self.lower_initial_builtin(),
+            // MLS §8.6: terminal() is false during ordinary simulation
+            // evaluation; terminal-event handling can override this phase
+            // marker explicitly when that event is implemented.
+            rumoca_core::BuiltinFunction::Terminal => Ok(self.emit_const(0.0)),
+            rumoca_core::BuiltinFunction::Sum => self.lower_sum_builtin(args, scope, call_depth),
+            rumoca_core::BuiltinFunction::Product => {
+                self.lower_product_builtin(args, scope, call_depth)
+            }
+            rumoca_core::BuiltinFunction::Size => self.lower_size_builtin(args, scope, call_depth),
+            rumoca_core::BuiltinFunction::Sample => {
+                self.lower_sample_builtin(args, scope, call_depth)
+            }
+            rumoca_core::BuiltinFunction::Reinit => Err(unsupported_at(
+                "reinit() must be converted to event update equations before Solve-IR row lowering",
+                span,
+            )),
+            _ => unreachable!("simple builtin handled before detailed lowering"),
+        }
+    }
+}
